+++ Common/src/main/java/fuzs/iteminteractions/common/api/v2/world/item/storage/ContainerStorage.java
@@ -196,7 +196,7 @@ public class ContainerStorage extends ComponentBackedStorage {
     @Override
     public TooltipComponent createTooltipImageComponent(ItemStack itemStack, Player player, NonNullList<ItemStack> itemList) {
         return new ItemContentsTooltip(itemList,
-                this.getSelectedItem(itemStack),
+                this.getSelectedItem(itemStack, player),
                 this.getGridWidth(itemList.size()),
                 this.getGridHeight(itemList.size()),
                 this.getResolvedDyeColor());