+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [v26.1.3-mc26.1.x] - 2026-06-10
 
+### Added
+
+- Add new `support_vanilla_connections` config option
+
 ### Fixed
 
 - Fix unable to use single item mode via scrolling while the cursor holds the item with contents